feat(api): Refactor trigger actions api to accept strings for type/target_type
This refactors the trigger actions endpoint to accept strings for type/target_type, which makes it
easier for frontend to work with.

We also refactor the registration of action handlers to help support this, and other things in the
future. We store a slug which we use to return from the endpoint, and also list the supported target
types for each handler. This will be used as part of returning a list of available actions to the
frontend as well in a future pr.